Move PreferencesContext
