+++ src/contexts/RootContextProvider.tsx
@@ -2,7 +2,7 @@ import { FC } from 'react';
 import { SnackbarProvider } from 'notistack';
 import { combineComponents } from '../utils/react';
 import UserProvider from '../features/user/contexts/UserContext';
-import PreferencesProvider from '../features/user/contexts/PreferencesContext';
+import PreferencesProvider from '../features/preferences/contexts/PreferencesContext';
 import BackdropProvider from './BackdropContext';
 
 const RootContextProvider = combineComponents(
