+++ src/features/genres/queries.ts
@@ -1,0 +1,10 @@
+import { gql } from '@apollo/client';
+
+export const QUERY_LIST_GENRES = gql`
+  query ListGenres {
+    listGenres {
+      externalId
+      name
+    }
+  }
+`;
